requirements.txt
Django>=5.0
psycopg2-binary>=2.9
djangorestframework>=3.14
django-cors-headers>=4.0
python-dotenv>=1.0
weasyprint==53.3
unidecode>=1.3
celery
redis
dj-database-url